customSearch/requirements.txt
Django == 1.9.6
djangorestframework == 3.3.3
elasticsearch == 2.3.0
psycopg2 == 2.6.1
